Moved and converted the api status components to ts
