+++ src/Components/ApiStatus/ApiErrorState.tsx
@@ -1,5 +1,4 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { FunctionComponent } from 'react';
 import {
   EmptyState,
   EmptyStateBody,
@@ -10,7 +9,11 @@ import {
 import { ExclamationCircleIcon } from '@patternfly/react-icons';
 import { global_danger_color_200 as globalDangerColor200 } from '@patternfly/react-tokens';
 
-const ApiErrorState = ({ message }) => (
+interface Props {
+  message: string | Record<string, any>;
+}
+
+const ApiErrorState: FunctionComponent<Props> = ({ message }) => (
   <EmptyState variant={EmptyStateVariant.small}>
     <EmptyStateIcon
       icon={ExclamationCircleIcon}
@@ -22,9 +25,5 @@ const ApiErrorState = ({ message }) => (
     <EmptyStateBody>{message}</EmptyStateBody>
   </EmptyState>
 );
-
-ApiErrorState.propTypes = {
-  message: PropTypes.string,
-};
 
 export default ApiErrorState;
