+++ src/Components/ApiStatus/NoData.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { FunctionComponent } from 'react';
 import {
   Title,
   EmptyState,
@@ -7,7 +7,7 @@ import {
 } from '@patternfly/react-core';
 import { CubesIcon } from '@patternfly/react-icons';
 
-const NoData = () => (
+const NoData: FunctionComponent<Record<string, never>> = () => (
   <EmptyState variant={EmptyStateVariant.full}>
     <EmptyStateIcon icon={CubesIcon} />
     <Title headingLevel="h5" size="lg">
