+++ src/Components/ModalContents.js
@@ -2,9 +2,9 @@ import React, { useCallback, useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import LoadingState from '../Components/LoadingState';
-import ApiErrorState from '../Components/ApiErrorState';
-import NoResults from './NoResults';
+import LoadingState from './ApiStatus/LoadingState';
+import ApiErrorState from './ApiStatus/ApiErrorState';
+import NoResults from './ApiStatus/NoResults';
 import Breakdown from '../Components/Breakdown';
 import JobStatus from '../Components/JobStatus';
 import { Paths } from '../paths';
