+++ src/Components/ModulesList.js
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { trimStr } from '../Utilities/helpers';
 import styled from 'styled-components';
-import LoadingState from '../Components/LoadingState';
-import NoData from '../Components/NoData';
+import LoadingState from './ApiStatus/LoadingState';
+import NoData from './ApiStatus/NoData';
 
 import {
   DataList,
