+++ src/Containers/Clusters/Clusters.js
@@ -2,7 +2,7 @@ import React, { useEffect, useCallback } from 'react';
 
 import { useQueryParams } from '../../Utilities/useQueryParams';
 
-import LoadingState from '../../Components/LoadingState';
+import LoadingState from '../../Components/ApiStatus/LoadingState';
 import {
   readClustersOptions,
   readJobExplorer,
@@ -30,7 +30,7 @@ import LineChart from '../../Charts/LineChart';
 import ModulesList from '../../Components/ModulesList';
 import TemplatesList from '../../Components/TemplatesList';
 import FilterableToolbar from '../../Components/Toolbar';
-import ApiErrorState from '../../Components/ApiErrorState';
+import ApiErrorState from '../../Components/ApiStatus/ApiErrorState';
 
 import { clusters } from '../../Utilities/constants';
 import useRequest from '../../Utilities/useRequest';
