+++ src/Containers/SavingsPlanner/Shared/Form/Steps/Templates/index.js
@@ -19,9 +19,9 @@ import {
   Td,
 } from '@patternfly/react-table';
 
-import LoadingState from '../../../../../../Components/LoadingState';
-import NoResults from '../../../../../../Components/NoResults';
-import ApiErrorState from '../../../../../../Components/ApiErrorState';
+import LoadingState from '../../../../../../Components/ApiStatus/LoadingState';
+import NoResults from '../../../../../../Components/ApiStatus/NoResults';
+import ApiErrorState from '../../../../../../Components/ApiStatus/ApiErrorState';
 import Pagination from '../../../../../../Components/Pagination';
 
 import { useQueryParams } from '../../../../../../Utilities/useQueryParams';
